fix(devtools): address diagnostics review feedback
- keep live-state diagnostics collection on the shared client lifecycle\n- separate metric and event subscriptions and cap retained log entries\n- normalize sparkline domains and remove the dead FPS aria label
